requirements.txt
anthropic==0.40.0
moviepy==1.0.3
requests==2.32.3
Pillow==10.4.0
numpy==1.26.4
python-dotenv==1.0.1
